Move the AgenticCheckoutSession to routes
